Removes native loader entrypoints from the SDK and thus from the public interface.
